Move case list models to case_list.py
Excludes case search models as these are conceptually separate.
